VOIP-7824: video: dsi: iproc: Support MIPI DSI 1080p
Change hs, hbp, hfp, vs, vbg, vfp from 8 bit to 16 bit. This allows
support of 1080p with values bigger than 256.

Reviewed-on: http://gerrit-ccxsw.broadcom.net/73522
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>

+++ drivers/video/fbdev/broadcom/bcm_iproc_dsi_ctrl.c
@@ -39,8 +39,8 @@ struct dsi_cfg {
 	bool ena_st_end;	/* enable ST_END */
 	bool ena_cnt_clk;	/* enable Continuous clock */
 	bool vmode;		/* 1 = Video Mode, 0 = Command Mode */
-	uint8_t hs, hbp, hfp, hbllp;
-	uint8_t vs, vbp, vfp;
+	uint16_t hs, hbp, hfp, hbllp;
+	uint16_t vs, vbp, vfp;
 };
 
 struct dsi_cmd {
